feat: 리뷰 평점 업데이트 구현

+++ backend/src/main/java/com/triply/barrierfreetrip/search/SearchDto.java
@@ -10,5 +10,5 @@ public class SearchDto {
     private String addr;
     private String tel;
     private String firstImage;
-    private Long rating;
+    private double rating;
 }
